Avoid code duplication moving the predefined ttfaf and cdf onto another scrit
